Really fix the session LRU manager
Turns out that my previous fix was again insufficient.  Looking into it
more, the problem is that the LRU manager is looking at the total memory
use, which means that when the server is doing heavy stuff (like a
running a submission) the amount of memory is huge, and the LRU thing is
degraded to just deducting a "life point" every memory poll, resulting
in the extremely short sessions that students see.

Fix this by creating a custodian around the web server, and using the
memory it holds for the accounting (which is done by copying and
tweaking `make-threshold-LRU-manager`).  With this, it's OK to lower the
memory limit to 1mb, which looks like it's more than enough (I see the
numbers get to around 90k).